2 fixes in chartJs models

+++ ChartExample - tutorial/ChartExample/Models/Dataset.cs
@@ -3,7 +3,7 @@
     public class Dataset
     {
         public string label { get; set; }
-        public int[] data { get; set; }
+        public double[] data { get; set; }
         public string[] backgroundColor { get; set; }
         public string[] borderColor { get; set; }
         public int borderWidth { get; set; }
